modularised code to just blogtemplate file...all blogs will follow same style
