+++ src/views/Blogs/LifeBlogs/Perseverance.vue
@@ -1,79 +1,17 @@
 <template>
-  <v-container class="blog-background">
-    <div class="font-weight-light text-h2 text-center">
-      Perseverance
-      <v-btn
-        v-show="!edit"
-        class="mx-2"
-        fab
-        dark
-        small
-        color="light-black"
-        @click="edit = !edit"
-      >
-        <v-icon dark> mdi-pencil </v-icon>
-      </v-btn>
-      <v-btn
-        v-show="edit"
-        class="mx-2"
-        fab
-        dark
-        small
-        color="light-black"
-        @click="
-          edit = !edit;
-          saveToDb();
-        "
-      >
-        <v-icon dark> mdi-content-save </v-icon>
-      </v-btn>
-    </div>
-
-    <ckeditor
-      v-if="edit"
-      :editor="editor"
-      v-model="editorData"
-      :config="editorConfig"
-    ></ckeditor>
-    <div
-      class="font-weight-light text-h6 text-left"
-      v-else
-      v-html="editorData"
-    ></div>
-  </v-container>
+  <BlogTemplate blogType="lifeblogs" :blogName="$options.name"></BlogTemplate>
 </template>
 
 <script>
-import ClassicEditor from "@ckeditor/ckeditor5-build-classic";
-import { getBlog, updateBlog } from "../../../firebase.js";
+import BlogTemplate from "../../../components/BlogTemplate.vue";
 
 export default {
   name: "Perseverance",
-  data: () => ({
-    edit: false,
-    editor: ClassicEditor,
-    editorData: "",
-    editorConfig: {
-      // The configuration of the editor.
-    },
-  }),
-  async mounted() {
-    this.editorData = await getBlog("lifeblogs", this.$options.name);
-  },
-  methods: {
-    async saveToDb() {
-      await updateBlog("lifeblogs", this.$options.name, {
-        value: this.editorData,
-      });
-    },
+  components: {
+    BlogTemplate,
   },
 };
 </script>
 
-
 <style>
-.blog-background {
-  background: url("https://vipul43.sirv.com/vipulsabout/dot-grid.png") repeat;
-  height: 100vh;
-}
 </style>
